fixed outputting format
fixed delete account to update properly

working on creating an initail account to avoid user entering n from the start. works until i enter lastname  to see customer info. thowing an error that there is an issue with the string but at the moment i have cin/cout being ambigous so that might be a contributing factor.